created new DayPage to help reduce repeated code
